Split up the servlets into 2 different files for clarity. After submitting a comment, changed the redirect to redirect to the comment section of the page.
