+++ portfolio/src/main/webapp/index.html
@@ -112,7 +112,7 @@
 
       <section id="comment-section" class="white-background">
         <div id="comment-content" class="section-content">
-          <form id="comment-form" action="/comment" method="POST">
+          <form id="comment-form" action="/add-comment" method="POST">
             <input type="text" name="name" placeholder="Name">
             <input type="text" name="message" placeholder="Drop a message!">
             <input id="comment-submit-btn" type="submit" class="purple-button" id="Submit">
